+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [catalog] App `file-manager.offspot.kiwix.org` to replace Edupi
+- [builder] `add_file()`'s `to` can now use `${APP_DIR:<ident>}` to target any app's home
+
+### Removed
+
+- [catalog] Removed `edupi.offspot.kiwix.org` package. See `file-manager.offspot.kiwix.org`
+
 ## [1.6.0] - 2024-01-08
 
 ### Changed